requirements.txt
selenium
notifypy
python-dotenv
